pyproject.toml
[project]
name = "talk-to-your-data-lab"
version = "0.1.0"
description = "GenBI chat PoC — natural language to SQL over Postgres"
readme = "PLAN.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "sqlglot>=26.0",
    "faker>=30.0",
    "pydantic>=2.9",
    "python-dotenv>=1.0",
    "typer>=0.24.1",
    "rich>=15.0.0",
]

[project.optional-dependencies]
ui = [
    "streamlit>=1.40",
    "plotly>=5.24",
    "pandas>=2.2",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/genbi"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests", "app", "evals"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "PL"]
ignore = ["PLR0913"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
